+++ src/__swaps__/screens/Swap/components/TokenList/TokenToSellList.tsx
@@ -1,35 +1,21 @@
-import c from 'chroma-js';
-import React, { useCallback, useMemo } from 'react';
-import { Text as RNText, StyleSheet } from 'react-native';
+import React, { useCallback } from 'react';
+import { StyleSheet } from 'react-native';
 import { CoinRow } from '@/__swaps__/screens/Swap/components/CoinRow';
 import { useAssetsToSell } from '@/__swaps__/screens/Swap/hooks/useAssetsToSell';
 import { ParsedSearchAsset } from '@/__swaps__/types/assets';
-import { Box, Stack, Text, Inline, Bleed, useColorMode, globalColors, HitSlop } from '@/design-system';
+import { Stack } from '@/design-system';
 import Animated, { runOnUI } from 'react-native-reanimated';
 import { useSwapContext } from '@/__swaps__/screens/Swap/providers/swap-provider';
 import { parseSearchAsset, isSameAsset } from '@/__swaps__/utils/assets';
-import { opacity } from '@/__swaps__/utils/swaps';
-import { useAccountAccentColor } from '@/hooks';
-import { ButtonPressAnimation } from '@/components/animations';
 import { ListEmpty } from '@/__swaps__/screens/Swap/components/TokenList/ListEmpty';
 import { FlashList } from '@shopify/flash-list';
+import { ChainSelection } from './ChainSelection';
 
 const AnimatedFlashListComponent = Animated.createAnimatedComponent(FlashList<ParsedSearchAsset>);
 
 export const TokenToSellList = () => {
-  const { accentColor: accountColor } = useAccountAccentColor();
-  const { isDarkMode } = useColorMode();
   const { SwapInputController } = useSwapContext();
   const userAssets = useAssetsToSell();
-
-  const accentColor = useMemo(() => {
-    if (c.contrast(accountColor, isDarkMode ? '#191A1C' : globalColors.white100) < (isDarkMode ? 2.125 : 1.5)) {
-      const shiftedColor = isDarkMode ? c(accountColor).brighten(1).saturate(0.5).css() : c(accountColor).darken(0.5).saturate(0.5).css();
-      return shiftedColor;
-    } else {
-      return accountColor;
-    }
-  }, [accountColor, isDarkMode]);
 
   const handleSelectToken = useCallback(
     (token: ParsedSearchAsset) => {
@@ -47,49 +33,8 @@ export const TokenToSellList = () => {
 
   return (
     <Stack space="20px">
-      <Box paddingHorizontal={'20px'}>
-        <Inline alignHorizontal="justify" alignVertical="center">
-          <Inline alignVertical="center" space="6px">
-            <Bleed vertical="4px">
-              <Box alignItems="center" justifyContent="center" width={{ custom: 18 }}>
-                <Bleed space={isDarkMode ? '16px' : undefined}>
-                  <RNText
-                    style={
-                      isDarkMode
-                        ? [
-                            styles.textIconGlow,
-                            {
-                              textShadowColor: opacity(accentColor, 0.2),
-                            },
-                          ]
-                        : undefined
-                    }
-                  >
-                    <Text align="center" color={{ custom: accentColor }} size="icon 13px" weight="black">
-                      􀣽
-                    </Text>
-                  </RNText>
-                </Bleed>
-              </Box>
-            </Bleed>
-            <Text color="label" size="15pt" weight="heavy">
-              My tokens
-            </Text>
-          </Inline>
-          <ButtonPressAnimation>
-            <HitSlop space="10px">
-              <Inline alignVertical="center" space="6px" wrap={false}>
-                <Text align="right" color={isDarkMode ? 'labelSecondary' : 'label'} size="15pt" weight="heavy">
-                  All Networks
-                </Text>
-                <Text align="center" color={isDarkMode ? 'labelTertiary' : 'labelSecondary'} size="icon 13px" weight="bold">
-                  􀆏
-                </Text>
-              </Inline>
-            </HitSlop>
-          </ButtonPressAnimation>
-        </Inline>
-      </Box>
+      <ChainSelection allText="All Networks" output={false} />
+
       <AnimatedFlashListComponent
         data={userAssets}
         ListEmptyComponent={<ListEmpty />}
